4/7/2020 Alex Dang
Fixed most of the problems with the home page. Made button changing color functionality. Made, but did not complete bulletin page. Fixed the navigation bar a bit. Fixed the royal mess that was colors.xml. Changed the layout to more closely match the design. Increased size of nav bar buttons and fixed their shape a little bit. Made the nav bar actually functional for the home page and bulletin page.